Server side functional - still a few bugs to fix

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface AccountDao {
     List<Account> getListOfAccounts();
     Account getOneAccount(long accountID);
-    BigDecimal getBalance(long accountID);
+    Double getBalance(long accountID);
     // get all transfers for one specific user
     List<Transfer> getListOfTransfers(long accountID);
 }
